+++ NEWS.md
@@ -25,6 +25,7 @@
 - Assert that `prework` is a language object, list of language objects, or character vector (#1 at pat-s/multicore-debugging on GitHub, @pat-s).
 - Use an environment instead of a list for `config$layout`. Supports internal modifications by reference. Required for #685.
 - Clean up the code of the parallel backends.
+- Make `dynamic` a formal argument of `target()`.
 
 
 # Version 7.7.0